+++ imio/dms/mail/migrations/migrate_to_3_1_6.py
@@ -13,6 +13,9 @@ class Migrate_To_3_1_6(Migrate_To_3_1):  # noqa
     def run_parts(self):
 
         if self.is_in_part("c"):
+            # Update d-print model
+            self.update_print_template()
+            # Correct localroles config
             self.fix_localroles_json_quoting()
             # update settings
             self.runProfileSteps("imio.dms.mail", steps=["plone.app.registry"])