+++ imio/dms/mail/tests/test_adapters.py
@@ -1592,6 +1592,7 @@ class TestOMApprovalAdapter(unittest.TestCase, ImioTestHelpers):
         self.assertFalse(pdf_file.signed)
         self.assertTrue(pdf_file.to_approve)
         self.assertTrue(pdf_file.approved)
+        self.assertFalse(pdf_file.to_print)
         self.assertEqual(pdf_file.content_category, "plone-annexes_types_-_outgoing_dms_files_-_outgoing-dms-file")
         self.assertFalse(hasattr(pdf_file, "conv_from_uid"))
 
@@ -1621,6 +1622,8 @@ class TestOMApprovalAdapter(unittest.TestCase, ImioTestHelpers):
         self.assertFalse(pdf_file.signed)
         self.assertTrue(pdf_file.to_approve)
         self.assertTrue(pdf_file.approved)
+        self.assertFalse(pdf_file.to_print)
+        self.assertFalse(self.files[1].to_print)
         self.assertEqual(pdf_file.content_category, "plone-annexes_types_-_outgoing_dms_files_-_outgoing-dms-file")
         self.assertFalse(hasattr(pdf_file, "conv_from_uid"))
 
@@ -1674,6 +1677,7 @@ class TestOMApprovalAdapter(unittest.TestCase, ImioTestHelpers):
         self.assertFalse(pdf_file.signed)
         self.assertTrue(pdf_file.to_approve)
         self.assertTrue(pdf_file.approved)
+        self.assertFalse(pdf_file.to_print)
         self.assertEqual(pdf_file.content_category, "plone-annexes_types_-_outgoing_dms_files_-_outgoing-dms-file")
         self.assertFalse(hasattr(pdf_file, "conv_from_uid"))